move TrimMobsSubConfig into Config class
